task1/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
pydantic==2.8.2
httpx==0.27.0
python-dotenv==1.0.1
